% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bees}
\alias{bees}
\title{Bee data}
\format{
## `bees`
A data frame with 20 rows and 14 columns:
\describe{
 \item{bee}{Unique identifier for each individual bee.}
 \item{condition}{Factor identifying the two experimental conditions. In the frustrated condition bees were temporarily restricted from returning to the hive after collecting nectar, in the free condition bees were able to return to the hive without delay.}
 \item{t1-t12}{Time between visits to the mechanical flower (in seconds) in each of 12 consecutive trials.}
}
}
\source{
Grice, J. W., Craig, D. P. A., & Abramson, C. I. (2015). A Simple and Transparent Alternative to Repeated Measures ANOVA. SAGE Open, 5(3), 215824401560419. https://doi.org/10.1177/2158244015604192
}
\usage{
bees
}
\description{
Time between visits to a mechanical flower by bees in two experimental conditions.
}
\keyword{datasets}
