% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_comparison.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Calculates PCCs and c-values based on pairwise comparison of conditions.}
\usage{
compare_conditions(
  result,
  cval_method = "exact",
  nreps = 1000L,
  progress = FALSE
)
}
\arguments{
\item{result}{an object of class "opafit" produced by a call to opa().}

\item{cval_method}{a string, either "exact" or "stochastic}

\item{nreps}{an integer, ignored if \code{cval_method = "exact"}}

\item{progress}{a boolean indicating whether to display a progress bar}
}
\value{
\code{compare_conditions} returns a list with the following elements

\describe{
  \item{pccs}{An upper triangle matrix containing PCCs calculated from each
  pairing of data columns, indicated by the matrix row and column names.}
  \item{cvals}{An upper triangle matrix containing c-values calculated from
  each pairing of data columns, indicated by the matrix row and column
  names.}
  }
}
\description{
Calculates PCCs and c-values based on pairwise comparison of conditions.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11),
                  t4 = c(10, 5, 11, 12))
opamod <- opa(dat, 1:4)
compare_conditions(opamod)
}
