% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cval_plot}
\alias{cval_plot}
\title{Plot individual chance values}
\usage{
cval_plot(m, threshold = NULL, title = TRUE, legend = TRUE)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
No return value, called for side effects.
}
\description{
Plot individual chance values
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
cval_plot(opamod)
cval_plot(opamod, threshold = 0.1)
}
