% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize}
\usage{
discretize(Input, N)
}
\arguments{
\item{Input}{vector or matrix of numeric values representing some continuously-valued landscape parameter}

\item{N}{The number of points to be created on the landscape}
}
\value{
Vector or matrix of the same dimensions as \code{Input}, but consisting of \emph{N} 1s and the remainder 0s
}
\description{
Converts a landscape of continuous-valued points into a discrete set of \emph{N} points
}
\examples{
X <- matrix(rnorm(10),nrow=10,ncol=1)
discretize(Input=X,N=5)

}
\seealso{
\code{\link{K.obj}},\code{\link{random.landscape}}
}

