% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{group_pcc}
\alias{group_pcc}
\title{Return the group PCC of the specified model}
\usage{
group_pcc(m)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
a double
}
\description{
Return the group PCC of the specified model
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
group_pcc(opamod)
}
