% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{group_pccs}
\alias{group_pccs}
\title{Return the group PCCs of the specified model}
\usage{
group_pccs(m)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
a numeric vector
}
\description{
Return the group PCCs of the specified model
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
group_pccs(opamod)
}
