% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{group_results.opafit}
\alias{group_results.opafit}
\title{Returns group-level PCC and chance values.}
\usage{
\method{group_results}{opafit}(m, digits = 2)
}
\arguments{
\item{m}{an object of class "opafit" produced by \code{opa()}.}

\item{digits}{a positive integer.}
}
\description{
Returns group-level PCC and chance values.
}
\details{
If the model was fitted with no grouping variable, a single PCC and c-value
are returned. If a grouping variable was specified in the call to \code{opa}
then PCCs and c-values are returned for each factor level of the grouping
variable.
}
\examples{
\dontrun{group_results(fitted_model)}
\dontrun{group_results(fitted_model, digits = 3)}
}
