% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hypothesis}
\alias{hypothesis}
\title{create a hypothesis object}
\usage{
hypothesis(xs, type = "pairwise")
}
\arguments{
\item{xs}{a numeric vector}

\item{type}{a string}
}
\value{
a list containing the following elements
}
\description{
create a hypothesis object
}
