% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{incorrect_pairs}
\alias{incorrect_pairs}
\title{Return the number of pairs of observations not matched by the hypothesis}
\usage{
incorrect_pairs(m)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
a non-negative integer
}
\description{
Return the number of pairs of observations not matched by the hypothesis
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
incorrect_pairs(opamod)
}
