% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{individual_cvals}
\alias{individual_cvals}
\title{Return the individual chance values of the specified model}
\usage{
individual_cvals(m)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
a numeric vector
}
\description{
Return the individual chance values of the specified model
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
individual_cvals(opamod)
}
