% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{individual_results.opafit}
\alias{individual_results.opafit}
\title{Returns individual-level PCC and chance values. If the model was fitted with
no grouping variable, a matrix of PCCs and c-values are returned
corresponding to the order of rows in the data. If the \code{opa} model was
fitted with a grouping variable specified, a table of PCCs and c-values is
returned ordered by factor level of the grouping variable.}
\usage{
\method{individual_results}{opafit}(m, digits = 2)
}
\arguments{
\item{m}{an object of class "opafit" produced by \code{opa()}}

\item{digits}{an integer}
}
\description{
Returns individual-level PCC and chance values. If the model was fitted with
no grouping variable, a matrix of PCCs and c-values are returned
corresponding to the order of rows in the data. If the \code{opa} model was
fitted with a grouping variable specified, a table of PCCs and c-values is
returned ordered by factor level of the grouping variable.
}
\examples{
\dontrun{individual_results(fitted_model)}
\dontrun{individual_results(fitted_model, digits = 3)}
}
