% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.obj.R
\name{K.obj}
\alias{K.obj}
\title{Object pattern analysis}
\usage{
K.obj(X, Type, Eps = NULL, r = NULL, which.marks = 1)
}
\arguments{
\item{X}{an object of class \code{"ppp"} containing \code{x} and \code{y} coordinates & object radii in the \code{marks} field.}

\item{Type}{character string indicating the shape of the landscape. Must be either \code{'rect'} or \code{'circle'}.}

\item{Eps}{the width of the ring used in the ring pattern calculation. It works best if it is quite small relative to the radii of the data and the sample radii \code{r}. It will be automatically calculated if not provided.
Specifying \code{Eps = 0} will coerce the code to perform regular point-pattern Ripley's \eqn{K} calculation, ignoring the object radii.}

\item{r}{distances at which to calculate \eqn{K}. The default will choose reasonable values.}

\item{which.marks}{optional argument; if the \code{ppp} object has a \code{marks} field with multiple columns, this indicates which column contains the object radii}
}
\value{
An object of class \code{"fv"}, similar to that created by \code{\link[spatstat]{Kest}}. 
       It is essentially a data frame with columns
       \itemize{
       \item \code{r}. the vector of values of the argument \code{r} at which \eqn{K_OR} has been estimated
       \item \code{K}. the vector of \eqn{K} or \eqn{K_OR} values (depending on the options selected) at each value of \code{r}
       \item \code{theo} the theoretical expected value of \eqn{K} ot \eqn{K_OR} under complete spatial randomnes
       \item \code{L} the value of \eqn{K} ot \eqn{K_OR} rescaled by \code{theo} to have expectation of 1 under complete spatial randomness.
       }
       
       Values of \code{L} > 1 indicate clustering; values < 1 indicate regularity. Note that the value of \code{L} corresonding to the smallest 
       value of \code{r} is often >> 1 just because it lies interior to some of the data objects.
}
\description{
Calculate \eqn{K_OR}: object-pattern equivalents to the point-pattern Ripley's \eqn{K} summary statistic
}
\details{
This function uses input and output objects in a way similar to package \code{spatstat}, and is specifically analogous to 
         \code{\link[spatstat]{Kest}}. To compute simulation envelopes for the \eqn{K} or \eqn{L} function, use \code{\link[spatstat]{envelope}} in the spatstat package.
         If \code{Eps = 0}, the function will compute the standard point-pattern analysis (PPA) Ripley's \eqn{K}. This is the same calculation as in 
         \code{\link[spatstat]{Kest}}, although that function has a wider range of edge-correction options and is much faster for PPA.
         If all of the radii in \code{X} are equal to zero, this function will also calculate the PPA version of \eqn{K}, but if \code{Eps > 0} or
         is \code{NULL}, the 'ring function' version of \eqn{K} will be calculated, as suggested by Wiegand & Molony (2004).
}
\examples{
X = random.landscape()
K.obj(X,Type='rect',Eps=0.1)

}
\author{
Will White, \email{whitejw@uncw.edu}
}
\references{
\href{http://onlinelibrary.wiley.com/doi/10.1111/j.0030-1299.2004.12497.x/full}{Wiegand and Moloney 2004}, Diegnan LK, Pawlik JR, Gleason ACR, White JW. "Object Pattern Analysis: an extension of Ripley's \eqn{K} to two-dimensional objects." Submitted, Methods in Ecology & Evolution
}
\seealso{
\code{\link{sim.autocorr}},  \code{\link{random.landscape}}, \code{\link[spatstat]{spatstat::Kest}}, \code{\link[spatstat]{spatstat::ppp.object}}
}

