% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no.overlap.R
\name{no.overlap}
\alias{no.overlap}
\title{No overlap}
\usage{
no.overlap(Input, R, Oa = 0)
}
\arguments{
\item{Input}{vector or matrix of binary values (1 and 0) representing the centroids of regularly-spaced object pattern.}

\item{R}{vector or matrix of the same dimension as Input, containing the radii of each object.}

\item{Oa}{an optional value specifying the proportional overlap allowed. Must be between 0 (default) and 1, inclusive.}
}
\value{
Vector or matrix of the same dimensions and number of points as Input, but with overlapping objects removed.
}
\description{
Filters an object-pattern landscape to eliminate overlapping objects
}
\examples{
X = matrix(round(runif(10)),nrow=5,ncol=2) ## create a simple object pattern
R = matrix(rnorm(10),nrow=5,ncol=2) ## Associated radii
no.overlap(Input=X,R=R,Oa=0)

}
\seealso{
\code{\link{K.obj}},\code{\link{random.landscape}}
}

