% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisperse.R
\name{overdisperse}
\alias{overdisperse}
\title{Overdisperse}
\usage{
overdisperse(Input, Mindist = 0)
}
\arguments{
\item{Input}{vector or matrix of binary values (1 and 0) representing a regularly spaced point pattern}

\item{Mindist}{the minimum distance allowed between points}
}
\value{
A vector or matrix of the same dimensions and number of points as \code{Input}, but rearranged so that no points are closer than \code{Mindist}
}
\description{
Cause a landscape of \emph{N} points to become overdispersed, with a minimum distance between them
}
\examples{
X = round(runif(20))
overdisperse(Input=X,Mindist=2)

}
\seealso{
\code{\link{K.obj}},\code{\link{random.landscape}},\code{\link{discretize}}
}

