% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pcc_plot}
\alias{pcc_plot}
\title{Plot individual PCCs.}
\usage{
pcc_plot(m, threshold = NULL, title = TRUE, legend = TRUE)
}
\arguments{
\item{m}{an object of class "opafit"}

\item{threshold}{a boolean indicating whether to plot a threshold abline}

\item{title}{a boolean indicating whether to include a plot title}

\item{legend}{a boolean indicating whether to include a legend}
}
\value{
No return value, called for side effects.
}
\description{
Plot individual PCCs.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
pcc_plot(opamod)
pcc_plot(opamod, threshold = 85)
}
