% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opa_condition_comparison}
\alias{plot.opa_condition_comparison}
\title{Plots PCCs and c-values computed from pairwise comparison of all conditions.}
\usage{
\method{plot}{opa_condition_comparison}(x, ...)
}
\arguments{
\item{x}{an object of class "opa_condition_comparison" produced by \code{compare_conditions()}}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Plots PCCs and c-values computed from pairwise comparison of all conditions.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
comp <- compare_conditions(opamod)
plot(comp)
}
