% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opa_hypothesis}
\alias{plot.opa_hypothesis}
\title{Plot a hypothesis.}
\usage{
\method{plot}{opa_hypothesis}(x, title = TRUE, ...)
}
\arguments{
\item{x}{an object of class "opa_hypothesis"}

\item{title}{a boolean indicating whether to include a plot title}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Plot a hypothesis.
}
\examples{
h <- hypothesis(c(1,2,3,3,3))
plot(h)
}
