% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opafit}
\alias{plot.opafit}
\title{Plots individual-level PCCs and chance-values.}
\usage{
\method{plot}{opafit}(x, ...)
}
\arguments{
\item{x}{an object of class "opafit" produced by \code{opa()}}

\item{...}{ignored}
}
\description{
Plots individual-level PCCs and chance-values.
}
\examples{
\dontrun{plot(fitted_model)}
}
