% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.oparandpccs}
\alias{plot.oparandpccs}
\title{Plot PCC replicates.}
\usage{
\method{plot}{oparandpccs}(x, ...)
}
\arguments{
\item{x}{an object of class "oparandpccs" produced by \code{random_pccs()}}

\item{...}{ignored}
}
\value{
no return value, called for side effects only.
}
\description{
Plot PCC replicates.
}
\details{
Plot a histogram of PCCs computed from randomly reordered data
used to calculate the chance-value.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
plot(random_pccs(opamod))
}
