% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.opa_hypothesis}
\alias{print.opa_hypothesis}
\title{Print details of a hypothesis}
\usage{
\method{print}{opa_hypothesis}(x, ...)
}
\arguments{
\item{x}{an object of type "opa_hypothesis"}

\item{...}{ignored}
}
\value{
No return value, called for side-effects.
}
\description{
Print details of a hypothesis
}
