% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.opafit}
\alias{print.opafit}
\title{Displays the call used to fit an ordinal pattern analysis model.}
\usage{
\method{print}{opafit}(x, ...)
}
\arguments{
\item{x}{an object of class "opafit".}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Displays the call used to fit an ordinal pattern analysis model.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
print(opamod)
}
