% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.opahypothesis}
\alias{print.opahypothesis}
\title{Print details of a hypothesis}
\usage{
\method{print}{opahypothesis}(x, ...)
}
\arguments{
\item{x}{an object of type "opaHypothesis"}

\item{...}{ignored}
}
\value{
No return value, called for side-effects.
}
\description{
Print details of a hypothesis
}
\examples{
h1 <- hypothesis(c(2, 1, 3, 4), type = "pairwise")
print(h1)
h2 <- hypothesis(c(2, 1, 3, 4), type = "adjacent")
print(h2)
}
