% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.pairwiseopafit}
\alias{print.pairwiseopafit}
\title{Displays the results of a pairwise ordinal pattern analysis.}
\usage{
\method{print}{pairwiseopafit}(x, ...)
}
\arguments{
\item{x}{an object of class "pairwiseopafit".}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Displays the results of a pairwise ordinal pattern analysis.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
pw <- compare_conditions(opamod)
print(pw)
print(pw, digits = 2)
}
