% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.landscape.R
\name{random.landscape}
\alias{random.landscape}
\title{Random landscape generation}
\usage{
random.landscape(units = 10, n = 10, res = 100, range.pts = 1,
  range.radii = 1, radius.mean = NULL, radius.sd = 1, Type = "rect")
}
\arguments{
\item{units}{the dimensions of the landscape (width of a square or diameter of a circle) in the desired units for x and y. For example, \code{units = 10} will produce \code{x} and \code{y} values that range from 0 to 10.}

\item{n}{number of points in the landscape}

\item{res}{the resolution of the landscape lattice. This is width (in number of lattice points) of the square study area, or diameter of the circular study area. A value ≥ 100 is recommended to approximate a continuous non-lattice landscape}

\item{range.pts}{the target autocorrelation scale of the point pattern data. A value greater than 0 and less than or equal to 1 will produce 
a random distribution (the actual value does not matter if it falls within that range). A value > 1 
produces an autocorrelated pattern with the specified autocorrelation range. A value < 0 produces an underdispersed 
pattern with that minimum distance between points (objects). A value of 0 produces an error.}

\item{range.radii}{the target autocorrelation scale of the object radii. This has the same usage as range.points, 
but acts on the object radii instead of the object centroids So values between 0 and 1 assign radii at random; values > 1 produce objects 
with radii that are spatially autocorrelated, and values < 0 produce a pattern in which objects repel each 
other at a scale equal to range.radii/2 beyond the boundary of their object
edge. In other words there will be a minimum distance of range.radii between the edges of any object.}

\item{radius.mean}{the mean of the object radii. Be advised that if values are too large then very few points will be
able to fit on the landscape.}

\item{radius.sd}{the standard deviation of the object radii. Be advised that if the standard deviation is very large relative to
the mean then many objects will be constrained to have radius 0 and the radii will have a truncated size distribution.}

\item{Type}{character string indicating the shape of the study area. Must be either \code{'rect'} (default) or \code{'circle'}.}
}
\value{
An object of class \code{"ppp"} containing coordinates for each object in fields \code{x} and \code{y}. Object radii are in the field \code{marks}
}
\description{
Create a landscape with various types of point patterns: complete spatial randomness, overdispersion, or underdispersion.
Can also create a landscape of objects, if object radii are provided. Uses FFT method based on Keitt (2000).
}
\examples{
random.landscape()

}
\author{
Will White, \email{whitejw@uncw.edu}
}
\references{
\href{http://link.springer.com/article/10.1023/A:1008193015770}{Keitt (2000)}
}
\seealso{
\code{\link{sim.autocorr}},  \code{\link{K.obj}}, \code{\link[spatstat]{ppp.object}}
}

