% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{random_pccs}
\alias{random_pccs}
\title{Return the random order generated PCCs used to calculate the group chance value}
\usage{
random_pccs(m)
}
\arguments{
\item{m}{an object of class "opafit"}
}
\value{
a numeric vector
}
\description{
Return the random order generated PCCs used to calculate the group chance value
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
random_pccs(opamod)
}
