% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.autocorr.R
\name{sim.autocorr}
\alias{sim.autocorr}
\title{Simulate autocorrelated landscape}
\usage{
sim.autocorr(n = 10, mean = 0, sd = 1, range = 1, dimension = 1)
}
\arguments{
\item{n}{length of spatial domain (points will be spaced equally). For 1-D this is the length; for 2-D this is the dimension of one side of the domain.}

\item{mean}{the spatial mean value for the landscape parameter}

\item{sd}{the standard deviation of landscape values}

\item{range}{the desired range of the landscape variogram. That is, the decorrelation scale.}

\item{dimension}{either 1 (for 1-D landscape) or 2 (for 2-D)}
}
\value{
Vector or matrix of dimension \code{n} containing the landscape values for each point.
}
\description{
Simulate a 1- or 2-dimensional landscape of autocorrelated values.
Uses FFT method based on Keitt (2000).
}
\examples{
sim.autocorr()

}
\references{
\href{http://link.springer.com/article/10.1023/A:1008193015770}{Keitt (2000)}
}
\seealso{
\code{\link{random.landscape}}
}

