% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summary.opafit}
\alias{summary.opafit}
\title{Prints a summary of results from a fitted ordinal pattern analysis model.}
\usage{
\method{summary}{opafit}(object, ..., digits = 2L)
}
\arguments{
\item{object}{an object of class "opafit".}

\item{...}{ignored}

\item{digits}{an integer used for rounding values in the output.}
}
\description{
Prints a summary of results from a fitted ordinal pattern analysis model.
}
\examples{
\dontrun{summary(opa_model)}
\dontrun{summary(opa_model, digits = 3)}
}
