% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hypothesis}
\alias{hypothesis}
\title{Create a hypothesis object}
\usage{
hypothesis(xs, type = "pairwise")
}
\arguments{
\item{xs}{a numeric vector}

\item{type}{a string}
}
\value{
a list containing the following elements
}
\description{
Create a hypothesis object
}
\examples{
h1 <- hypothesis(c(2, 1, 3, 4), type = "pairwise")
h2 <- hypothesis(c(2, 1, 3, 4), type = "adjacent")
}
