% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitopa.R
\name{opa}
\alias{opa}
\title{Fit an ordinal pattern analysis model}
\usage{
opa(
  dat,
  hypothesis,
  group = NULL,
  pairing_type = "pairwise",
  diff_threshold = 0,
  cval_method = "stochastic",
  nreps = 1000
)
}
\arguments{
\item{dat}{a data.frame}

\item{hypothesis}{a numeric vector}

\item{pairing_type}{a string}

\item{diff_threshold}{a positive integer or floating point number}
}
\value{
an object of class "opafit"
}
\description{
Fit an ordinal pattern analysis model
}
\details{
Determines the degree of match between data and a specified hypothesis.
Data is expected in \strong{wide} format with 1 row per individual and 1
column per measurement condition. Data must contain only columns consisting
of numerical values of the \emph{dependent} variable.

Any \emph{independent} variable must be passed separately as a vector with the
\code{group} keyword. The grouping vector must be a \emph{factor}.

The length of the \code{hypothesis} must be equal to the number of columns in
the dependent variable data.frame \code{dat}.
}
\examples{
\dontrun{opa(dat, c(1,2,3,3))}
\dontrun{opa(dat, c(1,2,3,3), pairing_type = "adjacent")}
\dontrun{opa(dat, c(1,2,3,3), diff_threshold = 1)}
\dontrun{opa(dat, c(1,2,3,3), group = group_vector)}
}
\references{
Grice, J. W., Craig, D. P. A., & Abramson, C. I. (2015). A Simple and Transparent Alternative to Repeated Measures ANOVA. SAGE Open, 5(3), 215824401560419. <https://doi.org/10.1177/2158244015604192>
Thorngate, W. (1987). Ordinal Pattern Analysis: A Method for Assessing Theory-Data Fit. Advances in Psychology, 40, 345–364. <https://doi.org/10.1016/S0166-4115(08)60083-7>
}
