% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pcc_threshold_plot}
\alias{pcc_threshold_plot}
\title{Plots individual PCCs relative to a user-supplied PCC threshold value.}
\usage{
pcc_threshold_plot(m, pcc_threshold = 75)
}
\arguments{
\item{m}{an object of class "opafit"}

\item{pcc_threshold}{a numeric scalar}
}
\value{
an object of class "ggplot"
}
\description{
Plots individual PCCs relative to a user-supplied PCC threshold value.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
pcc_threshold_plot(opamod)
pcc_threshold_plot(opamod, pcc_threshold = 85)
}
