% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pituitary}
\alias{pituitary}
\title{Childhood growth data}
\format{
## `pituitary`
A data frame with 108 rows and 4 columns:
\describe{
 \item{distance}{distance in milimetres from the pituitary to the pteryo-maxillary fissure}
 \item{age}{age in years}
 \item{individual}{identifier for each individual}
 \item{sex}{sex of each individual}
}
}
\source{
Potthoff, R. F., & Roy, S. N. (1964). A Generalized Multivariate Analysis of Variance Model Useful Especially for Growth Curve Problems. Biometrika, 51(3/4), 313–326. https://doi.org/10.2307/2334137
}
\usage{
pituitary
}
\description{
Distance from the the pituitary to the pteryo-maxillary fissure.
}
\keyword{datasets}
