% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opafit}
\alias{plot.opafit}
\title{Plots individual-level PCCs and chance-values.}
\usage{
\method{plot}{opafit}(x, ...)
}
\arguments{
\item{x}{an object of class "opafit" produced by \code{opa()}}

\item{...}{ignored}
}
\value{
an object of class "ggplot"
}
\description{
Plots individual-level PCCs and chance-values.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
opamod <- opa(dat, 1:3)
plot(opamod)
}
