% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opafit}
\alias{plot.opafit}
\title{Plots individual-level PCCs and chance-values.}
\usage{
\method{plot}{opafit}(x, pcc_threshold = NULL, cval_threshold = NULL, ...)
}
\arguments{
\item{x}{an object of class "opafit" produced by \code{opa()}}

\item{pcc_threshold}{a number used as the x-intercept to plot a PCC threshold abline}

\item{cval_threshold}{a number used as the x-intercept to plot a c-value threshold abline}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Plots individual-level PCCs and chance-values.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
plot(opamod)
}
