% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.opafit}
\alias{plot.opafit}
\title{Plots individual-level PCCs and chance-values.}
\usage{
\method{plot}{opafit}(m)
}
\arguments{
\item{m}{an object of class "opafit" produced by \code{opa()}}
}
\description{
Plots individual-level PCCs and chance-values.
}
\examples{
\dontrun{plot(fitted_model)}
}
