% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_hypothesis}
\alias{plot_hypothesis}
\title{Plot a hypothesis.}
\usage{
plot_hypothesis(h, title = TRUE)
}
\arguments{
\item{h}{a numeric vector}

\item{title}{a boolean indicating whether to include a plot title}
}
\value{
No return value, called for side effects.
}
\description{
Plot a hypothesis.
}
\examples{
h <- c(1,2,3,3,3)
plot_hypothesis(h)
}
