% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_hypothesis}
\alias{plot_hypothesis}
\title{Plot a hypothesis.}
\usage{
plot_hypothesis(h)
}
\arguments{
\item{h}{a numeric vector}
}
\value{
an object of class "ggplot"
}
\description{
Plot a hypothesis.
}
\examples{
h <- c(1,2,3,3,3)
plot_hypothesis(h)
}
