% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summary.opafit}
\alias{summary.opafit}
\title{Prints a summary of results from a fitted ordinal pattern analysis model.}
\usage{
\method{summary}{opafit}(object, ..., digits = 2L)
}
\arguments{
\item{object}{an object of class "opafit".}

\item{...}{ignored}

\item{digits}{an integer used for rounding values in the output.}
}
\value{
No return value, called for side effects.
}
\description{
Prints a summary of results from a fitted ordinal pattern analysis model.
}
\examples{
dat <- data.frame(t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
h <- hypothesis(1:3)
opamod <- opa(dat, h)
summary(opamod)
summary(opamod, digits = 3)
}
